Playground for designing the front end of the site
As we are unsure how to implement any of the CCS, JS and Bootstrap
the mockup folder has been created so we can get the design and layout
implemented without having to worry about any of the wagtail complexities.

Once we have a draft for the design we will be able to incorporate
this into the wagtail templates.